+++ core/memory/__init__.py
@@ -1,3 +1,3 @@
-﻿from core.memory.memory_manager import MemoryManager
+from core.memory.memory_manager import MemoryManager
 
 __all__ = ['MemoryManager']