+++ tools/search/__init__.py
@@ -1,3 +1,3 @@
-﻿from tools.search.web_search_tool import WebSearchTool
+from tools.search.web_search_tool import WebSearchTool
 
 __all__ = ['WebSearchTool']