+++ tools/video/__init__.py
@@ -1,3 +1,3 @@
-﻿from tools.video.ffmpeg_tool import FFmpegTool
+from tools.video.ffmpeg_tool import FFmpegTool
 
 __all__ = ['FFmpegTool']